+++ REXCompiler/omptests/rex-simd/benchmark/matmul/rose/rex_lib_matmul_float_pf.c
@@ -38,7 +38,7 @@ void OUT__1__9606__matmul_simd__34__(int *__global_tid,int *__bound_tid)
     for (__index_ = __lower_; __index_ <= __upper_; __index_ += 1) {
       for (_p_j = ((size_t )0); _p_j < ((unsigned long )1024); _p_j++) {
         float temp = (float )0;
-        svbool_t __pg0 = svwhilelt_b32((unsigned long )0,((unsigned long )1024) - 1);
+        svbool_t __pg0 = svwhilelt_b32((unsigned long )0,(unsigned long )(((unsigned long )1024) - 1));
         svfloat32_t __part0 = svdup_f32(0.00000L);
         __pg0 = svptrue_b32();
         temp += svaddv(__pg0,__part0);
