+++ REXCompiler/omptests/rex-simd/benchmark/matmul/rose/rose_matmul_float_unroll_sve.c
@@ -37,7 +37,7 @@ void matmul_simd(float **A,float **B,float **C)
     for (size_t j = 0; j < 1024; j++) {
       temp = 0;
       size_t k = 0;
-      svbool_t __pg0 = svwhilelt_b32((unsigned long )0,((unsigned long )1024) - 1);
+      svbool_t __pg0 = svwhilelt_b32((unsigned long )0,(unsigned long )(((unsigned long )1024) - 1));
       svfloat32_t __part0 = svdup_f32(0.00000L);
       svfloat32_t __part7 = svdup_f32(0.00000L);
       svfloat32_t __part14 = svdup_f32(0.00000L);
@@ -46,7 +46,7 @@ void matmul_simd(float **A,float **B,float **C)
       svfloat32_t __part35 = svdup_f32(0.00000L);
       svfloat32_t __part42 = svdup_f32(0.00000L);
       svfloat32_t __part49 = svdup_f32(0.00000L);
-      for (k = 0; k <= ((unsigned long )1024) - 1; k += 8 * svcntw()) {
+      for (k = 0; k <= (((unsigned long )1024) - 1); k += 8 * svcntw()) {
         float *__ptr1 = A[i];
         svfloat32_t __vec2 = svld1(__pg0,&__ptr1[k]);
         float *__ptr3 = B[j];
@@ -103,7 +103,7 @@ void matmul_simd(float **A,float **B,float **C)
         svfloat32_t __vec54 = svmul_f32_m(__pg0,__vec53,__vec51);
         svfloat32_t __vec55 = svadd_f32_m(__pg0,__vec54,__part49);
         __part49 = (__vec55);
-        __pg0 = svwhilelt_b32(k,((unsigned long )1024) - 1);
+        __pg0 = svwhilelt_b32(k,(unsigned long )(((unsigned long )1024) - 1));
       }
       __pg0 = svptrue_b32();
       temp += svaddv(__pg0,__part49);
