+++ REXCompiler/omptests/rex-simd/benchmark/sum/rose/rose_sum_float_unroll_sve.c
@@ -30,7 +30,7 @@ float sum(float *X)
 {
   size_t i;
   float result = 0;
-  svbool_t __pg0 = svwhilelt_b32((unsigned long )0,((unsigned long )10240) - 1);
+  svbool_t __pg0 = svwhilelt_b32((unsigned long )0,(unsigned long )(((unsigned long )10240) - 1));
   svfloat32_t __part0 = svdup_f32(0.00000L);
   svfloat32_t __part3 = svdup_f32(0.00000L);
   svfloat32_t __part6 = svdup_f32(0.00000L);
@@ -39,7 +39,7 @@ float sum(float *X)
   svfloat32_t __part15 = svdup_f32(0.00000L);
   svfloat32_t __part18 = svdup_f32(0.00000L);
   svfloat32_t __part21 = svdup_f32(0.00000L);
-  for (i = ((size_t )0); i <= ((unsigned long )10240) - 1; i += 8 * svcntw()) {
+  for (i = ((size_t )0); i <= (((unsigned long )10240) - 1); i += 8 * svcntw()) {
     svfloat32_t __vec1 = svld1(__pg0,&X[i]);
     svfloat32_t __vec2 = svadd_f32_m(__pg0,__vec1,__part0);
     __part0 = (__vec2);
@@ -64,7 +64,7 @@ float sum(float *X)
     svfloat32_t __vec22 = svld1(__pg0,&X[i + svcntw() * 7]);
     svfloat32_t __vec23 = svadd_f32_m(__pg0,__vec22,__part21);
     __part21 = (__vec23);
-    __pg0 = svwhilelt_b32(i,((unsigned long )10240) - 1);
+    __pg0 = svwhilelt_b32(i,(unsigned long )(((unsigned long )10240) - 1));
   }
   __pg0 = svptrue_b32();
   result += svaddv(__pg0,__part21);
